+++ src/shaders/transposematmul.wgsl
@@ -12,9 +12,9 @@ struct Params {
   output_offset: u32,
 };
 
-@binding(0) @group(0) var<storage, read_write> output : array<f16>;
-@binding(1) @group(0) var<storage, read> left : array<f16>;
-@binding(2) @group(0) var<storage, read> right : array<f16>;
+@binding(0) @group(0) var<storage, read_write> output : array<f32>;
+@binding(1) @group(0) var<storage, read> left : array<f32>;
+@binding(2) @group(0) var<storage, read> right : array<f32>;
 @binding(3) @group(0) var<uniform> params : Params;
 
 const WORKGROUP_SIZE_X = 8;
